requirements.txt
requests>=2.31.0,<3
openai>=1.30.0,<2
resend>=2.0.0,<3
python-dotenv>=1.0.0,<2
msal>=1.28.0,<2

